+++ REQUIREMENTS.md
@@ -17,7 +17,7 @@
       - [x] 선택된 상품이 없을 때 버튼 비활성화
       - [x] 선택된 상품이 있을 때 버튼을 누르면 선택된 상품이 장바구니에서 삭제된다.
     - [x] 상품 상세 목록
-      - [ ] 헤더에 장바구니에 담긴 상품 개수를 표시한다.
+      - [x] 헤더에 장바구니에 담긴 상품 개수를 표시한다.
       - [x] 상품 목록 아이템
         - [x] 상품 선택 버튼, 상품 이미지, 상품 이름, 상품삭제버튼, 수량 선택버튼, 상품 가격을 표시한다.
         - [x] 상품 선택 버튼을 클릭 해서 상품을 선택할 수 있다.
