+++ src/constants/actionType.js
@@ -1,0 +1,14 @@
+const ACTION_TYPE = {
+  SET_ITEM_LIST: 'SET_ITEM_LIST',
+
+  ADD_ITEM_TO_CART: 'ADD_ITEM_TO_CART',
+  SET_CART_ITEM_LIST: 'SET_CART_ITEM_LIST',
+  TOGGLE_CART_ITEM_CHECKBOX: 'TOGGLE_CART_ITEM_CHECKBOX',
+  SET_ALL_CART_ITEM_CHECKBOX: 'SET_ALL_CART_ITEM_CHECKBOX',
+  SET_CART_ITEM_QUANTITY: 'SET_CART_ITEM_QUANTITY',
+  DELETE_CART_ITEMS: 'DELETE_CART_ITEMS',
+
+  ADD_ORDER_DETAIL: 'ADD_ORDER_DETAIL',
+};
+
+export default ACTION_TYPE;
