+++ src/constants/index.js
@@ -2,3 +2,4 @@ export { default as COLOR } from './color';
 export { default as PATH } from './path';
 export { default as BASE_URL } from './baseURL';
 export { default as MESSAGE } from './message';
+export { default as ACTION_TYPE } from './actionType';
